Move common func to lensUtil
