+++ api/v1/controllers/lenses.js
@@ -19,6 +19,7 @@ const httpStatus = require('../constants').httpStatus;
 const apiErrors = require('../apiErrors');
 const AdmZip = require('adm-zip');
 const redisCache = require('../../../cache/redisCache').client;
+const lensUtil = require('../../../utils/lensUtil');
 
 const ZERO = 0;
 const ONE = 1;
@@ -110,65 +111,6 @@ function handleLensMetadata(requestObj, libraryParam, seqObj) {
   }
 }
 
-/**
- * Create Json for files in Lens library with format, filename: file contents.
- * @param  {Object} lensObject - lens object
- * @returns {Object} zipContents - Json prepared from library contents.
- */
-function createLensLibraryJson(lensObject) {
-  const re = /(?:\.([^.]+))?$/;
-  const zip = AdmZip(lensObject.library);
-  const zipEntries = zip.getEntries();
-
-  const zipContents = {};
-  for (let j = 0; j < zipEntries.length; j++) {
-    let ext = re.exec(zipEntries[j].entryName)[ONE] || '';
-    ext = ext.toLowerCase();
-
-    // handle different file types here
-    if (ext === 'png' || ext === 'jpg' || ext === 'jpeg') {
-      const b64str = zipEntries[j].getData().toString('base64');
-      zipContents[zipEntries[j].entryName] = b64str;
-    } else {
-      zipContents[zipEntries[j].entryName] = zip.readAsText(
-        zipEntries[j]
-      );
-    }
-  }
-
-  return zipContents;
-}
-
-/**
- * Recursively cleans the object (i.e. calls "get" on any sequelize
- * instances), strips out nulls (because swagger validation doesn't like
- * nulls).
- *
- * @param {Object} obj - The object to clean
- * @returns {Object} - The cleaned object
- */
-function cleanAndCreateLensJson(obj) {
-  const o = obj.get ? obj.get({ plain: true }) : obj;
-  if (o) {
-    const keys = Object.keys(o);
-    for (let i = 0; i < keys.length; i++) {
-      const key = keys[i];
-
-      if (key === 'library' && o[key]) {
-        const lensLibraryJson = createLensLibraryJson(o);
-        o[key] = lensLibraryJson;
-      } else if (o[key] === undefined || o[key] === null) {
-        delete o[key];
-      } else if (Array.isArray(o[key])) {
-        o[key] = o[key].map((j) => cleanAndCreateLensJson(j));
-      } else if (typeof o[key] === 'object') {
-        o[key] = cleanAndCreateLensJson(o[key]);
-      }
-    }
-  }
-
-  return o;
-}
 
 /**
  * Prepares the object to be sent back in the response ("cleans" the object,
@@ -183,7 +125,7 @@ function cleanAndCreateLensJson(obj) {
  *  the response
  */
 function responsify(rec, props, method) {
-  const o = cleanAndCreateLensJson(rec);
+  const o = lensUtil.cleanAndCreateLensJson(rec);
   o.apiLinks = u.getApiLinks(o.id, props, method);
   return o;
 }
@@ -404,6 +346,4 @@ module.exports = {
     })
     .catch((err) => u.handleError(next, err, helper.modelName));
   },
-  cleanAndCreateLensJson,
-
 }; // exports
